psbt: Remove redundant pre-iteration of data when deserializing
Instead, have the various elements allocated and added on demand using the
existing functions. This is substantially simpler and allows us to
remove redundant checks during parsing that the add/set functions already
perform.

As part of this change, some parameter constification was required.
Also, wally_psbt_set_global_tx is updated to not create empty/invalid inputs
and outputs.