psbt: check for segwit utxos before non-segwit when signing
This mimics the logic where no non-segwit utxo was given, i.e. in core
prior to 20.1, by ignoring it when a segwit one is present.

TODO: The signing logic needs to be fixed to recognise segwit inputs
independently of which type of utxo is passed (its perfectly valid to
only pass the non segwit utxo, although core thankfully does not do this
yet to avoid breaking implementations like ours).