psbt: fix some errors on add/remove input/output failure
The addition of new allocations for v2 inputs and outputs means the
error handling logic must be updated. Some new checks are needed for v2
which in the v0 path are caught by the tx modification path.

Additionally, there was a long-existing bug in the old code where memory
allocation failure would cause the number of in/outputs to be off-by-one.