psbt: allow fetching prev txid, output index and sequence from v0 PSBTs
Although these fields can only be set on PSBT inputs in v2, in v0 they
are available on the global tx, so allow the getters to return them in
this case to make the interface more consistent (and better for signing,
since these fields are needed there).

Also fix the invalid parameter checks for these calls.