psbt: verify any provided sighash matches sigs when parsing
Core doesn't verify this on parsing, BIP-0174 states only that
finalizers must fail to finalize where sighashes don't match.

Block this from happening when parsing to avoid processing an invalid
PSBT (that a buggy finalizer may not catch), and to avoid needing to
check the sighash in the finalization code for each supported input
type. Note we already prevent signatures being added that don't match.