descriptor: remove erroneous top-level kind check
If WALLY_MINISCRIPT_ONLY is not passed, our descriptor can be mixed
(contain both descriptor and miniscript expressions).

The caller can check if a descriptor is pure (contains no miniscript)
using the feature flag WALLY_MS_IS_DESCRIPTOR if they do not want to
allow mixed expressions.

Add a test case proving that mixed expressions can be parsed with either
flags (using pk() which is valid as miniscript and a descriptor
expression).