psbt: allow setting a global tx with scriptSigs and witnesses
Rather than forcing the caller to remove these before setting the tx, ignore
them when cloning into the global tx. Note that any global tx in a serialized
PSBT v0 must still not contain these tx input elements as this is
invalid according to BIP-174.

(Strictly speaking, witnesses should never be present in the global tx
because it should be serialized without them. An upcoming change adds a
test case for this).