descriptor-policy: add a flag for enforcing unique keypaths in policies
Also fix and test a few edge cases in ensuring substitution invariants.

The relevant BIP makes this mandatory, however given it is somewhat
expensive to verify, and the security concerns mentioned seem to only
relate to miniscript malleation via old tx signatures, make it opt-in.

It would be nice to support ensuring this for non-policy miniscript
descriptors. But, given the combination of possible descriptor key types
and path expressions I do not believe it is feasible for any implementation
to do this correctly without exactly the other limits that policies
enforce. A general solution would have to solve all possible paths and derive
all combinations of allowed keys which is not trivial to prove correct and
extremely compute intensive and thus not feasible on HWW.