build: rename base58/64.c and hex.c to avoid build issues
Based on a patch by Marko Bencun <marko@shiftcrypto.ch>.

Original commit description/rationale:

When configuring libwally core with these flags:

    ./configure --enable-static --disable-shared

and compiling using `make`, libtool outputs this message:

    copying selected object files to avoid basename conflicts...

It copies the base64 and hex object files as they have the same
basename as the files built from `src/ccan/ccan/base64/base64.c`
`ccan/ccan/str/hex/hex.c`.

By using `make V=1` one can see that it is trying to copy it using
`ln` (a hardlink). This does not work properly inside a Docker
container on macOS, where the hard link is created but cannot be used
immediately ("file not found" in the next step where the object files
are used in linking). We tested this on two macOS machines, and it
only consistently did not work on one of them. We could not figure out
the root cause for the error. Giving the files unique names avoids the
copying/hardlinking and avoids the error.