build: remove obsolete wheel building scripts, update internal CI
wally wheels should be built using standard tools, e.g.

'build': 'python -m build" to build an sdist and then build a wheel for that sdist.
'pip': 'pip install .' to build a wheel and install it.

Now that standard builds are supported, update the internal CI
docker image/config to build an isolated wheel from its sdist and
smoke test it. The tested sdist can then be uploaded to PyPI to
enable pip installs for systems we don't provide binaries for.