descriptor: support parsing descriptors as elements descriptors
Elements core supports standard descriptors but generates (at least for
taproot) different scriptpubkeys/addresses since it uses a different hash
to tweak its internal keys. Allow specifying that a descriptor is an
Elements descriptor when parsing, and use this to perform the correct
tweak.

Its unclear whether other descriptor expressions are affected since
there is no documentation of this behaviour.

Elements (only via rust-elements) also supports a renamed set of
descriptors and extensions which we do not attempt to support in this
change.