descriptor: add support for elements el-prefixed descriptor builtins
These are not documented anywhere and are only implemented in
rust-miniscript, which renames descriptor builtins to be prefixed
with "el" - making them incompatible with any descriptor currently
supported by Elements core.

However the test vectors from ELIP-0150 include elsh(wpkh()) expressions
in addition to elwpkh() expressions, which seems to indicate that mixing
prefixed and non-prefixed builtins should be accepted.

This implementation therefore allows mixing prefixed and non-prefixed
builtins, with the descriptor being marked as Elements if either a
prefixed builtin is given, or if the descriptor is explicitly parsed as
Elements via WALLY_MINISCRIPT_AS_ELEMENTS.